Keep the on_border flag
If we are too eager to generate the new id, i.e. find out the root
of the agglomeration result (s in the original code), the on_border
flag will be remove after it, and any domain sharing the same root
appearing in the later part of the loop will be mapped to 0.