Try harder when agglomerating chunked data
It is not necessary to stop the agglomeration when we hit a single
frozen supervoxel. We only needs to skip the edge when it can be reached
by two frozen chain descending from the boundary.

Because of the implementation, we can no longer using the disjoint set
and let the code figure out the representative supervoxel. We have to
use the frozen supervoxel to represent the segment if there is one, thus
we can recognize them in the future step of the agglomeration. So it is
impossible to use the segment with a smaller number of neighbours when
we update the incident matrix. On the brighter side we no longer need
the disjoint set which saves some memory.