Reduce the size of the ongoing remaps
There are many domains maps to the same intermediate domains before the
chunks are completely merged already. Instead of storing all of them in
the ongoing remaps, let us pick the first domain encountered in each
atomic chunk as the representative to save in the ongoing remaps, while
marking the reset of the domains to the representative as done.

We will need to sort out the appropriate final supervoxel ids like we
do in the mean affinity agglomeration process. This might take some time
and need to be benchmarked