Switch to unordered_map
The reason I choose ordinary map before is it is impossible to delete
elements in an unordered_map while we are looping over its elements.
However it is better to just delete the whole map after we finish the
loop than update its content all the time, since all the elements will
be merged into other maps anyway.