Replace map with vector part 1
Instead of using a map of maps incident matrix to store neighbors of
each supervoxel, it is faster and more memory efficient to use a map of
sorted vectors.

When we load all the region graph, we sorted them according to the two
segid specifying the edge, this implicitly change the way we resolve the
degenerate cases. The result is identical when using double precision.