Init the supervoxel sizes to 1
The first encounters of most of the supervoxels happen during region
graph process. We originally initialize its size to 0 and update it to
the actual value when we meet them again reading the supervoxel sizes.
This approach does not work for atomic chunks, which have empty
supervoxels sizes and rely on the region graph information to initialize
all segments sizes to 1. So now I change the approach to initialize the
supervoxel sizes to 1 when processing the region graph