Agglomerate more edges
It is not necessary to freeze any edges contain only one frozen node.
Only when the other node is also frozen or it has frozen neighbours.
Additionally, since we apply the size dependent heuristics, we also need
to delay the agglomeration if we cannot determine whether we should
agglomerate it based on the information available, i.e. One of the node
have more than 1000 supervoxels and it is not frozen. In principle we
can do a little bit more by rejecting edges already satisfy the
heuristics even when some of the nodes are frozen, but probably not
worth it.