Keep track of the segments in other chunks
It is not enough to record all the boundary supervoxels processed in the
current chunk, we have to include all the supervoxels merged through
those boundary supervoxels and merge them when we match the chunks
before agglomerate the rest of the supervoxels. otherwise in the case
the "path" of the agglomeration makes a U-turn, the results will be
different.

Add a new entry in the output to distinguish the supervoxels truly
located at the boundary and the supervoxels included through those
boundary supervoxels