Freeze segments if they touch a frozen segment
In the non-overlapped mode it is possible to make an extra shortcut,
namely only the segments touching more than two frozen segments. In that
case we change the order of the agglomeration but not the final results.
However this will require non-trivial merges after overlapped
agglomeration. Keep it simple for now and hope the speed-up of the
overlapped mode can still be significant