Ignore irrelevant segments in the remap file
Because of the changes in the reduce_chunk.py in the script repo, it is
possible the remap input contains segments completely outside of the
current chunk. Allow the code to continue if it encountered some segment
that is neither done nor ongoing. This is not as safe as the old
behavior, but I do not know a better solution for now