Use the whole cut plane to keep track of the overlapped region
The old code determines which partial result is relevant for a
particular direction by associating it with the boundary segments
crossing the center core cube. This is not enough when the segment make
a 90 degree turn or even more complicated maneuver in the overlapped
region. We have to use the cut plane crossing the "super chunk" to
determine whether to keep the information for the particular direction