Replace the segid unconditionally
Because we make sure all boundaries shared the same agglomeration
results when we agglomerate with overlap, it is impossible to have the
unmerged boundaries containing more advanced remap than the ones we
merged at a particular level. So update the chunk boundary no matter
what the sizes.

The real issue this change fixes is the scenario when two different
partial results has the same number of supervoxels. This can happen
after we allow agglomerating extra segments completely contained in the
current chunk.