Incorporate semantic labels
the current approach is simply prevent merger between different
categories, i.e. axons cannot merge with dendrites. This rule no longer
apply when we consider complete neurons.

In order to have the overlapping mode working with semantic labels, we
need to actively veto edges that have already been considered in the
overlapping step. The dominant semantic label of the segment can change
as it absorbs more and more small piece. This does not necessarily give
us the right answer, but it is consistent.