Fix the check for extra agglomerations
The purpose of the check is to make sure the heuristics, size-dependent
cuts and the semantic cuts will always be applied in proper condition.
The extra agglomeration can agglomerate a pair of segments even if one
of them is incomplete, but the size-dependent cuts and semantic cuts
need the whole segment to complete to make the final decision.