Optimize remap_edges and merge_edges
We arrange the rg_vector in such a way that edges actively
participated in the agglomeration is in front. So remap_edges only
need to deal with the inactive part of rg_vector. Similarly we
only need to merge edges in the later half of the rg_vector, for
the active edges we only check if they are nil.