Use the dominant semantic label
For each chunked supervoxels, collect all semantic labels and use the
dominant one to represent the supervoxels. In principle we can do better
and pick the dominant label after chunked supervoxels are merged, right
now we only use the dominant label of the dominant chunked supervoxel.

The semantic label is stored in the first element of the array; The
second element is the number of voxels associated with this label; The
third one is the total number of voxels.