More flexible padding when cutting chunks
We do not extract all edges in a single atomic chunk in each atomic
task, instead we only extend the chunk in the forward direction and
collect edges in a chunk offset by the forward padding. To determine
incomplete edges, we still need to scan a larger chunk covering the
atomic chunk plus the long range affinites. The boundary supervoxels
need to be consistent between different chunks. When we reach the edges
of the dataset we add enough padding so all the boundary voxels are
padded 0 voxels